refactor: change directory name to kakao from oauth2
