+++ src/main/java/bitnagil/bitnagil_backend/auth/kakao/model/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package bitnagil.bitnagil_backend.auth.oauth2.model;
+package bitnagil.bitnagil_backend.auth.kakao.model;
 
 import java.util.Collection;
 import java.util.Map;
