+++ src/main/java/bitnagil/bitnagil_backend/auth/kakao/service/OAuth2TokenService.java
@@ -1,4 +1,4 @@
-package bitnagil.bitnagil_backend.auth.oauth2.service;
+package bitnagil.bitnagil_backend.auth.kakao.service;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -9,7 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoUserInfoResponse;
+import bitnagil.bitnagil_backend.auth.kakao.response.KakaoUserInfoResponse;
 
 /**
  * 카카오 OAuth2 인증을 위한 토큰 발급 및 사용자 정보 조회를 담당하는 서비스입니다.
