+++ src/main/java/bitnagil/bitnagil_backend/user/service/UserAuthService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import bitnagil.bitnagil_backend.auth.jwt.RefreshToken;
 import bitnagil.bitnagil_backend.auth.jwt.Token;
 import bitnagil.bitnagil_backend.auth.jwt.JwtProvider;
-import bitnagil.bitnagil_backend.auth.oauth2.service.OAuth2TokenService;
-import bitnagil.bitnagil_backend.auth.oauth2.service.RedisService;
+import bitnagil.bitnagil_backend.auth.kakao.service.OAuth2TokenService;
+import bitnagil.bitnagil_backend.auth.kakao.service.RedisService;
 import bitnagil.bitnagil_backend.global.errorcode.ErrorCode;
 import bitnagil.bitnagil_backend.global.exception.CustomException;
 import bitnagil.bitnagil_backend.user.Repository.UserRepository;
@@ -15,8 +15,8 @@ import bitnagil.bitnagil_backend.enums.SocialType;
 import bitnagil.bitnagil_backend.auth.jwt.TokenResponse;
 import bitnagil.bitnagil_backend.user.entity.User;
 import bitnagil.bitnagil_backend.enums.Role;
-import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoAccount;
-import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoUserInfoResponse;
+import bitnagil.bitnagil_backend.auth.kakao.response.KakaoAccount;
+import bitnagil.bitnagil_backend.auth.kakao.response.KakaoUserInfoResponse;
 import lombok.RequiredArgsConstructor;
 
 /**
